[COREV] Implement memcpy HWLP optimization in MachineInstr lowering
This implements a custom lowering function from the `CV_HWLP_MEMCPY`
pseudo instruction to the machine instructions building the resulting
hardware loop instructions.

To make sure, that instructions inside the HW loop don't get compressed,
this commit adds a field to the `RISCVMachineFunctionInfo`, where all
basic blocks used by the HW loop are stored. This is then used when
emitting the ELF and Asm files to prevent compression.

Signed-off-by: Philipp Krones <philipp.krones@embecosm.com>