[COREV] Correctly lower MAC intrinsics to the MAC instructions
One the one hand this deals with the type legalization for the builtins,
which partly have `i16` as input types. On the other hand it implements
custom lowering of those builtins to the instruction. The custom
lowering has to be used, because patterns in tablegen files are not
powerful enough to represent the type conversions.

Signed-off-by: Philipp Krones <philipp.krones@embecosm.com>