[COREV] Re-enable Windows testing
The commit that added CI for the COREV LLVM fork disabled Windows
testing due to timeouts, but this looks to be just a case of cold
build caches, and really could occur wiht any of the tests. This
re-enables the Windows tests, and explains the situation of what to
do should the tests be cancelled due to hitting the 360-minute limit.