[RISCV] Use macros as C API interfaces of CORE-V intrinsics with const arguments.
This patch adds macros in the C API header files instead of wrapped
calls for those intrinsics which use const arguments.
This change affects the intrinsics for:

SIMD:

- add_h
- sub_h
- cplxmul_r
- cplxmul_i
- subrotmj
- shuffle_xhi
- extract[u]_[h/b]
- insert_[h/b]

Bitmanip:

- bitrev

It also changes the type of the immediate argument of extract and shuffle.sci
to const in the clang intrinsics definitions.