[RISCV][Clang] Add C API header for bitmanip CORE-V intrinsics.
This commit adds the C API header to serve as an interface for
the intrinsic functions provided with the CORE-V xcvbitmanip
extension.

The commit includes the addition of the header itself, the update of
cmake to support this header file and more to follow from RISC-V
and a clang CodeGen test that checks the integrity of the call to the
wrapper functions designed in the C API CORE-V bitmanip header.