[RISCV][Clang] Fix types of CORE-V simd Clang builtins.
This patch fixes the types of simd CORE-V Clang builtins as per spec.
The patch also adds the `I` qualifier to the immediate argument of
extract[u]_[h,b] intrinsics so that clang errors when a variable is
passed instead of letting the compiler crash later.

The patch also updates the clang CoreGen test accordingly.