[RISCV][Clang] Add C API header for elw CORE-V intrinsics.
This commit adds the C API header to serve as an interface for
the intrinsic functions provided with the CORE-V xcvelw extension.

The commit includes the addition of the header itself and a clang
CodeGen test that checks the integrity of the call to the wrapper
functions designed in the C API CORE-V elw header.