[RISCV][CV-SIMD] Check value range of constant operands of simd builtins.
Add a check for the constant operands of SIMD clang builtins to
be in range.
This is added for the builtins: add.h, sub.h, shuffle.sci.h (the b
version uses all the bits of the constant operand and so no need for
a custom range check), cplxmul.r, cplxmul.i, subrotmj.

Add test cases for the corresponding out of range errors.
Use in range values in simd.c test for shuffle.sci.h.